separate out the markdown rewriting code
